Feat/delete verification

+++ src/modules/Livro/service/DeleteLivro.service.ts
@@ -1,3 +1,4 @@
+import { IExemplarRepository } from '@modules/Exemplar/repository/IExemplarRepository.interface';
 import { EntityNotFoundError } from '@shared/errors/EntityNotFoundError';
 import { inject, injectable } from 'tsyringe';
 import { ILivroRepository } from '../repository/ILivroRepository.interface';
@@ -7,6 +8,9 @@ class DeleteLivroService {
   constructor(
     @inject('LivroRepository')
     private livroRepository: ILivroRepository,
+
+    @inject('ExemplarRepository')
+    private exemplarRepository: IExemplarRepository,
   ) {}
 
   async execute(liv_Id: string) {
@@ -18,7 +22,17 @@ class DeleteLivroService {
       throw new EntityNotFoundError('Livro não encontrado tente novamente');
     }
 
-    // adicionar verificação de se existe exemplar
+    const exemplares = await this.exemplarRepository.listBy({
+      filter: {
+        exe_liv_id: livro.liv_Id,
+      },
+    });
+
+    if (exemplares.results.length > 0) {
+      throw new EntityNotFoundError(
+        'Livro possui exemplares impossível deletar',
+      );
+    }
 
     await this.livroRepository.delete(livro);
   }
